Renamed several moving pieces of Elastic.Channels to better reflect their responsibilities
