+++ src/Elastic.Ingest.Elasticsearch/DataStreams/DataStreamChannelOptions.cs
@@ -5,9 +5,9 @@ using Elastic.Transport;
 
 namespace Elastic.Ingest.Elasticsearch.DataStreams
 {
-	public class DataStreamChannelOptions<TEvent> : ElasticsearchChannelOptionsBase<TEvent>
+	public class DataStreamResponseItemsChannelOptions<TEvent> : ElasticsearchResponseItemsChannelOptionsBase<TEvent>
 	{
-		public DataStreamChannelOptions(HttpTransport transport) : base(transport) =>
+		public DataStreamResponseItemsChannelOptions(HttpTransport transport) : base(transport) =>
 			DataStream = new DataStreamName(typeof(TEvent).Name.ToLowerInvariant());
 
 		public DataStreamName DataStream { get; set; }
